backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.13
sqladmin==0.20.1
python-multipart==0.0.9
jinja2==3.1.4
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
email-validator==2.3.0
python-dotenv==1.0.1
faster-whisper==1.2.1
